BugFix: Reimplement getHostStatus using Socket API
Testing scheduled tasks, I found connections using telnet were failing most of the time with error "Error:Connection refused" . Debugging pointed to getHostStatus() that is called before connection to the device. It turns out our, getHostStatus is using fsockopen to verify connectivity, but it fails to close the socket. I tested adding fclose(fp), but it did not solve the problem. It turns out, sockets are not closed inmmediatley and a work around must be implemented as documented in user notes in http://php.net/manual/en/function.socket-close.php
I also fixed the comments.